Tweaking day 5 and adding day 6

+++ Test.AdventOfCode2020/TestDay05Functions.cs
@@ -14,7 +14,7 @@ namespace Test.AdventOfCode2020
         [TestCase("BBFFBBFRLL", 102, 4)]
         public void TestExamples(string fullString, long expectedRow, long expectedCol)
         {
-            Tuple<long, long> answer = Day05Functions.GetRowAndColumn(fullString);
+            Tuple<int, int> answer = Day05Functions.GetRowAndColumn(fullString);
             Assert.AreEqual(expectedRow, answer.Item1, "row");
             Assert.AreEqual(expectedCol, answer.Item2, "row");
         }
